+++ packages/signals_core/lib/src/core/signals.dart
@@ -192,6 +192,8 @@ abstract class _Listenable {
 
   void _notify();
 
+  void onDispose(EffectCleanup cleanup);
+
   void dispose();
 }