+++ packages/signals_core/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+
+- Adding `dispose` methods for `signal`, `computed` and `effect`
+- Adding `initialValue` getter for `signal` and `computed`
+
 ## 1.0.1
 
 - Exposing `reloadSignalsDevTools` for hot reload support